+++ src/CommunicationInterfaceWebServer.cpp
@@ -24,145 +24,8 @@
 #include <unistd.h>
 #include <cstring>
 
-
-bool exitNow = false;
-
-
 namespace XBot {
-
-bool HttpCivetHandler::handleGet(CivetServer *server, struct mg_connection *conn) {
-      
-      const char* uri =mg_get_request_info(conn)->request_uri;
-      std::string suri(uri);
-      http_interface->setUri(suri);
-      
-      const char* query = mg_get_request_info(conn)->query_string;
-      if (query!=nullptr){
-          std::string squery(query);
-          http_interface->setQuery(squery);
-	  std::string key = squery.substr(0, squery.find("="));
-	  std::string val = squery.substr(squery.find("=")+1);
-	  http_interface->setKey(key);
-	  http_interface->setVal(val);
-      }
-          
-      std::shared_ptr<ResponseInterface> resp_interface;
-      http_interface->handleGet(resp_interface);
-      std::string type = resp_interface->GetTypeResponse();
-      std::string header = "HTTP/1.1 200 OK\r\nContent-Type: "
-			    +type+"\r\nConnection: close\r\n\r\n";
-
-      mg_printf(conn, header.c_str());
-      mg_write(conn,resp_interface->GetData(),resp_interface->GetLength());
-  
-      return true;
-}  
-
-bool HttpCivetHandler::handlePost(CivetServer* server, mg_connection* conn){
-  
-      const char* uri =mg_get_request_info(conn)->request_uri;
-      std::string suri(uri);
-      http_interface->setUri(suri);
-      
-      const struct mg_request_info *req_info = mg_get_request_info(conn);
-      long long rlen = 0;
-      long long nlen = 0;
-      long long tlen = req_info->content_length;
-      //std::cout<<"tlen "<<tlen<<std::endl;
-      //NOTE fixed size
-      char buf[4096];
-  
-      memset(buf,0,4096);
-      while (nlen < tlen) {
-          rlen = tlen - nlen;
-          if (rlen > sizeof(buf)) {
-                  rlen = sizeof(buf);
-          }
-          rlen = mg_read(conn, buf, (size_t)rlen);
-          if (rlen <= 0) {
-                  break;
-          }
-         
-       }
-      
-      std::shared_ptr<RequestObject> req = std::make_shared<RequestObject>();
-      req->SetData(buf);
-      http_interface->handlePost(req);
-      
-      mg_printf(conn,"{\"response\":\"ok\"}");
-  
-      return true;
-}  
-  
-bool WebSocketHandler::handleConnection(CivetServer *server, const struct mg_connection *conn) {
-    
-    std::cout<<"WS connected\n";
-    return true;  
-}
-
-void WebSocketHandler::handleReadyState(CivetServer *server, struct mg_connection *conn) {
-   
-    std::cout<<"WS ready\n";
-    const char *text = "Hello from XBotCore";
-    mg_websocket_write(conn, WEBSOCKET_OPCODE_TEXT, text, strlen(text));    
-    sharedData->increaseNumClient();    
-} 
-        
-bool WebSocketHandler::handleData(CivetServer *server,
-                                struct mg_connection *conn,
-                                int bits,
-                                char *data,
-                                size_t data_len) {
-    
-  
-  
-    data[data_len] = 0;
-    std::shared_ptr<RequestObject> req = std::make_shared<RequestObject>();
-    req->SetData(data);
-    
-    //TODO move in abstraction interface
-    void * buff;
-    buff = req->GetData();     
-    StringStream stream((char*)buff);
-    //std::cout<<"pos"<<std::string((char*)buff)<<std::endl;
-    Document d;
-    d.ParseStream(stream);
-    
-    sharedData->clearJointMap();      
-    if( d.HasMember("joint")){          
-        assert(d["joint"].isArray());
-        const Value& array = d["joint"];
-        for (SizeType i = 0; i < array.Size(); i++){
-            const Value& obj = array[i];
-            int id = obj["id"].GetInt();
-            double val = obj["val"].GetDouble();
-            sharedData->insertJoint(id,val);              
-        }        
-      }    
-        
-    //read robot state
-    WebRobotState rstate;
-    bool resp = buffer->remove(rstate);
-        
-    StringBuffer buffer;
-    const char* btosend = nullptr;    
-    buffer.Clear();
-    if(resp){      
-        rstate.serialize(buffer);
-        btosend = buffer.GetString();
-    }
-        
-    if( btosend!=nullptr)
-        mg_websocket_write(conn, WEBSOCKET_OPCODE_TEXT, btosend, buffer.GetLength());
-        
-    return true;
-}  
-
-void WebSocketHandler::handleClose(CivetServer *server, const struct mg_connection *conn) {
-       std::cout<<"WS closed\n";
-       sharedData->decreaseNumClient();
-}        
-        
+              
 CommunicationInterfaceWebServer::CommunicationInterfaceWebServer():
     CommunicationInterface()
 {
