+++ src/Web/HttpHandler.cpp
@@ -1,0 +1,121 @@
+/*
+ * Copyright (C) 2017 IIT-ADVR
+ * Author:  Giuseppe Rigano
+ * email:   giuseppe.rigano@iit.it
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>
+*/
+
+#include <HttpHandler.h>
+
+HttpHandler::HttpHandler (std::shared_ptr<SharedData>& sharedData, std::shared_ptr<Buffer<WebRobotState>>& buffer){
+      
+      this->sharedData = sharedData;
+      this->buffer = buffer;
+}
+ 
+void HttpHandler::handleGet(std::shared_ptr<ResponseInterface>& response){      
+         
+      if(uri.compare("/switch")==0){
+        sharedData->insertSwitch(key, val);       
+      }
+      else if(uri.compare("/cmd")==0) {
+        sharedData->insertCmd(key, val);        
+      }       
+      else if(uri.compare("/webmaster")==0){
+        sharedData->setMaster(key);
+      }
+      
+      std::string sresp="";
+      sresp+="<html><body>\r\n";
+      sresp+=("<h2>XBOTCORE </h2>\r\n");
+      
+      for( auto const &s : sharedData->getAllStatus()){                 
+          auto const &outer_key = s.first;
+          auto const &inner_map = s.second;
+          std::string ss="<h3>"+ outer_key+ " "+ inner_map+ "</h3>\r\n";
+          const char * w =const_cast<char*>( ss.c_str());
+          sresp+=(w);
+      }
+
+      sresp+=("</body></html>\r\n");
+
+      response = std::make_shared<StringResponse>(sresp);
+     
+//       std::shared_ptr<StringBuffer> jsonresp = std::make_shared<StringBuffer>();
+//       // 1. Parse a JSON string into DOM.
+//       const char* json = "[{\"project\":\"rapidjson\",\"stars\":10}, {\"project\":\"rapidjson\",\"stars\":10}]";
+//       Document d;
+//       d.Parse(json);
+// 
+//       // 2. Modify it by DOM.
+//       Value& s = d["stars"];
+//       s.SetInt(s.GetInt() + 1);
+// 
+//       // 3. Stringify the DOM
+//       Writer<StringBuffer> writer(*jsonresp);
+//       d.Accept(writer);
+//       
+//       response = std::make_shared<JsonResponse>(jsonresp);
+     
+             
+}
+  
+void HttpHandler::handlePost(std::shared_ptr<RequestObject>& request){
+    
+      void * buff;
+      buff = request->GetData();     
+      StringStream stream((char*)buff);
+      //std::cout<<"pos"<<std::string((char*)buff)<<std::endl;
+      Document d;
+      d.ParseStream(stream);
+      
+      sharedData->clearJointMap();
+      
+      std::vector<double> vec;
+      if(uri.compare("/alljoints")==0){     
+        if( d.HasMember("link_position")){        
+          assert(d["link_position"].isArray());
+          const Value& array = d["link_position"];
+          for (SizeType i = 0; i < array.Size(); i++){
+              double val = array[i].GetDouble();
+              vec.push_back(val);   
+          }
+          sharedData->external_command->add(vec);
+          //HACK simulation of holding value for longer time
+          sharedData->external_command->add(vec);
+        }
+      }else if(uri.compare("/singlejoint")==0){
+        
+        //{"joint":[{"id": 15, "val": 0},{"id": 16, "val": 0}]}
+        if( d.HasMember("joint")){          
+          assert(d["joint"].isArray());
+          const Value& array = d["joint"];
+          for (SizeType i = 0; i < array.Size(); i++){
+              const Value& obj = array[i];
+              int id = obj["id"].GetInt();
+              double val = obj["val"].GetDouble();
+              sharedData->insertJoint(id,val);              
+          }        
+        }
+        
+      }
+            
+      /*StringBuffer buffer;
+      Writer<StringBuffer> writer(buffer);
+      d.Accept(writer);
+      std::cout <<"stringify"<< std::string(buffer.GetString()) << std::endl;*/ 
+    
+}
+
